Plotting_myself
Done from scratch, all other formatting is commented out. Defeats the purpose of this branch but it's 4:30am. Think the first 10 lines of he csv file are still being plotted which throws the axes off by a lot. Feels like I'm close though. This will work great with the MFA to create vowel plots once that's up and running.